+++ CHANGELOG.md
@@ -5,15 +5,19 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [Unreleased]
 
-## [0.4.0] - 2023-01-18
-
 ### Added
+
 - remove_unknowns xtor argument to drop entries with missing info
+- get_diagnoses() include_locations arg to return disorder location
 
 ### Fixed
 
 - diagnoses DF to exclude NA entries
 
+
+### Removed
+
+- KPDVDB.get_locations() method (merged with get_diagnoses())
 
 ## [0.3.1] - 2023-01-17